move the column renaming to data validation API
